requirements.txt
Django>=4.2,<5.0
django-jazzmin==2.6.0
django-ckeditor-5==0.2.13
Pillow>=10.0.0
python-decouple>=3.8
